% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_geboes_score}
\alias{assert_geboes_score}
\title{Check if a value is a valid Geboes score}
\usage{
assert_geboes_score(x, all_categories = FALSE)
}
\arguments{
\item{x}{A character vector of Geboes score grades}

\item{all_categories}{Must \code{x} include all categories exactly once?}
}
\value{
\code{x}, if the values are valid; an error otherwise
}
\description{
Check if a value is a valid Geboes score
}
