% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_geboes_score.R
\name{factor_geboes_score}
\alias{factor_geboes_score}
\alias{levels_geboes_score}
\title{Convert a Geboes score into a factor}
\usage{
factor_geboes_score(x, drop_0 = FALSE)

levels_geboes_score(drop_0 = FALSE)
}
\arguments{
\item{x}{A character vector of Geboes score grades}

\item{drop_0}{Drop the ".0" levels other than "0.0"}
}
\value{
\code{x} as a factor, verified to be a Geboes score
}
\description{
Convert a Geboes score into a factor
}
\section{Functions}{
\itemize{
\item \code{levels_geboes_score()}: Get all available levels of the Geboes score

}}
