% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geboes_score_continuous.R
\name{geboes_continuous}
\alias{geboes_continuous}
\title{Convert Geboes scores to a continuous measurement}
\usage{
geboes_continuous(x)
}
\arguments{
\item{x}{A character vector of Geboes score grades}
}
\value{
The numeric value of the continuous Geboes score (the sum of each of
the category/subcategory scores)
}
\description{
Convert Geboes scores to a continuous measurement
}
\references{
Zenlea T, Yee EU, Rosenberg L, et al. Histology Grade Is Independently
Associated With Relapse Risk in Patients With Ulcerative Colitis in Clinical
Remission: A Prospective Study. Official journal of the American College of
Gastroenterology | ACG. 2016;111(5):685. doi:10.1038/ajg.2016.50
}
