% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_geboes_score.R
\name{max_geboes_score}
\alias{max_geboes_score}
\title{Determine the maximum value of the Geboes score}
\usage{
max_geboes_score(x, na.rm = TRUE)
}
\arguments{
\item{x}{A character vector of Geboes score grades}

\item{na.rm}{a logical indicating whether missing values should be
    removed.}
}
\value{
The maximum value of \code{x} where values ending in ".0" are considered
as "0.0"
}
\description{
Determine the maximum value of the Geboes score
}
